chore: extract event bus to file
