+++ packages/@tinacms/core/src/events.test.ts
@@ -1,59 +1,22 @@
-type Callback = Function
-type Unsubscribe = Function
+/**
 
-interface CMSEvent {
-  type: string
-}
+Copyright 2019 Forestry.io Inc
 
-class Listener {
-  constructor(private callback: Callback, private events: string[] = []) {}
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
 
-  handleEvent(event: CMSEvent) {
-    if (this.watchesEvent(event)) {
-      this.callback(event)
-    }
-  }
+    http://www.apache.org/licenses/LICENSE-2.0
 
-  watchesEvent(currentEvent: CMSEvent) {
-    if (!this.events.length) return true
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
 
-    const watchCurrentEvent = !!this.events.find(watchedEvent => {
-      return matchEventPattern(currentEvent, watchedEvent)
-    })
+*/
 
-    return watchCurrentEvent
-  }
-}
-
-function matchEventPattern(event: CMSEvent, pattern: string) {
-  const eventParts = event.type.split(':')
-  const patternParts = pattern.split(':')
-
-  let index = 0
-  let ignoresEvent = false
-
-  while (!ignoresEvent && index < patternParts.length) {
-    const wildcard = patternParts[index] === '*'
-    const matchingParts = patternParts[index] === eventParts[index]
-    ignoresEvent = !(wildcard || matchingParts)
-    index++
-  }
-
-  return !ignoresEvent
-}
-
-class EventBus {
-  private listeners: Listener[] = []
-
-  subscribe(callback: Callback): Unsubscribe {
-    this.listeners.push(new Listener(callback))
-    return () => {}
-  }
-
-  dispatch(event: CMSEvent) {
-    this.listeners.forEach(listener => listener.handleEvent(event))
-  }
-}
+import { EventBus, Listener } from './event'
 
 describe('EventBus', () => {
   it('calls listener and passes it the dispatched event', () => {
